Revert "parallel tc"
This reverts commit 6abc7e547b789a88f8ed288927782c54f3c2fe79.

+++ assembler/src/include/io/kmer_iterator.hpp
@@ -17,7 +17,7 @@ raw_kmer_iterator<Seq> make_kmer_iterator(const std::string &FileName,
 
 template<class Seq>
 std::vector<raw_kmer_iterator<Seq>> make_kmer_iterator(const std::string &FileName,
-                                                       size_t K, size_t amount) {
+                                                       unsigned K, unsigned amount) {
   std::vector<raw_kmer_iterator<Seq>> res;  
   if (amount == 1) {
     res.emplace_back(FileName, Seq::GetDataSize(K));
